MQTT/mqtt-python/requirements.txt
awsiotsdk==1.5.17